Backend/requirements.txt
# FastAPI and Server
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.19
pydantic==2.10.3
pydantic-settings==2.6.1

# Image Processing
opencv-python==4.10.0.84
Pillow==11.0.0
numpy==1.26.4

# OCR - Google Cloud Vision API
google-cloud-vision==3.8.1

# NLP and Text Processing
nltk==3.9.1
spacy==3.8.2
# Run: python -m spacy download en_core_web_sm

# Embeddings and Similarity
sentence-transformers==3.3.1
scikit-learn==1.5.2

# Transformers for NLI
transformers==4.47.0
torch==2.5.1

# Enhanced AI Models
google-generativeai==0.8.3

# Configuration
PyYAML==6.0.2
python-dotenv==1.0.1

# CORS
fastapi-cors==0.0.6

# Logging and Monitoring
loguru==0.7.2

# Machine Learning (for model training)
xgboost==2.1.3
pandas==2.2.3

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
